Reorganizes NumericsHelpers into different files using partial classes.
