+++ Robust.Shared.Maths/NumericsHelpers.Divide.cs
@@ -1,0 +1,295 @@
+using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.Intrinsics;
+using System.Runtime.Intrinsics.Arm;
+using System.Runtime.Intrinsics.X86;
+
+namespace Robust.Shared.Maths
+{
+    public static unsafe partial class NumericsHelpers
+    {
+        #region Divide
+
+        /// <summary>
+        ///     Divide a by b and stores the result in a.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        public static void Divide(Span<float> a, ReadOnlySpan<float> b)
+        {
+            Divide(a, b, a);
+        }
+
+        /// <summary>
+        ///     Divide a by b and stores the result in s.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        public static void Divide(ReadOnlySpan<float> a, ReadOnlySpan<float> b, Span<float> s)
+        {
+            if (a.Length != b.Length || a.Length != s.Length)
+                throw new ArgumentException("Length of arrays must be the same!");
+
+            if (Enabled)
+            {
+                if (AvxEnabled && LengthValid256Single(a.Length))
+                {
+                    DivideAvx(a, b, s);
+                    return;
+                }
+
+                if (LengthValid128Single(a.Length))
+                {
+                    if (Sse.IsSupported)
+                    {
+                        DivideSse(a, b, s);
+                        return;
+                    }
+
+                    if (AdvSimd.Arm64.IsSupported)
+                    {
+                        DivideAdvSimd(a, b, s);
+                        return;
+                    }
+                }
+            }
+
+            DivideNaive(a, b, s, 0, a.Length);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideNaive(ReadOnlySpan<float> a, ReadOnlySpan<float> b, Span<float> s, int start, int end)
+        {
+            for (var i = start; i < end; i++)
+            {
+                s[i] = a[i] / b[i];
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideSse(ReadOnlySpan<float> a, ReadOnlySpan<float> b, Span<float> s)
+        {
+            var remainder = a.Length & 3;
+            var length = a.Length - remainder;
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrB = b)
+                {
+                    fixed (float* ptrS = s)
+                    {
+                        for (var i = 0; i < length; i += 4)
+                        {
+                            var j = Sse.LoadVector128(ptr + i);
+                            var k = Sse.LoadVector128(ptrB + i);
+
+                            Sse.Store(ptrS + i, Sse.Divide(j, k));
+                        }
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideAdvSimd(ReadOnlySpan<float> a, ReadOnlySpan<float> b, Span<float> s)
+        {
+            var remainder = a.Length & 3;
+            var length = a.Length - remainder;
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrB = b)
+                {
+                    fixed (float* ptrS = s)
+                    {
+                        for (var i = 0; i < length; i += 4)
+                        {
+                            var j = AdvSimd.LoadVector128(ptr + i);
+                            var k = AdvSimd.LoadVector128(ptrB + i);
+
+                            AdvSimd.Store(ptrS + i, AdvSimd.Arm64.Divide(j, k));
+                        }
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideAvx(ReadOnlySpan<float> a, ReadOnlySpan<float> b, Span<float> s)
+        {
+            var remainder = a.Length & 7;
+            var length = a.Length - remainder;
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrB = b)
+                {
+                    fixed (float* ptrS = s)
+                    {
+                        for (var i = 0; i < length; i += 8)
+                        {
+                            var j = Avx.LoadVector256(ptr + i);
+                            var k = Avx.LoadVector256(ptrB + i);
+
+                            Avx.Store(ptrS + i, Avx.Divide(j, k));
+                        }
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        #endregion
+
+        #region DivideByScalar
+
+        /// <summary>
+        ///     Divide a by scalar b and stores the result in a.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        public static void Divide(Span<float> a, float b)
+        {
+            Divide(a, b, a);
+        }
+
+        /// <summary>
+        ///     Divide a by scalar b and stores the result in s.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        public static void Divide(ReadOnlySpan<float> a, float b, Span<float> s)
+        {
+            if (a.Length != s.Length)
+                throw new ArgumentException("Length of arrays must be the same!");
+
+            if (Enabled)
+            {
+                if (AvxEnabled && LengthValid256Single(a.Length))
+                {
+                    DivideAvx(a, b, s);
+                    return;
+                }
+
+                if (LengthValid128Single(a.Length))
+                {
+                    if (Sse.IsSupported)
+                    {
+                        DivideSse(a, b, s);
+                        return;
+                    }
+
+                    if (AdvSimd.Arm64.IsSupported)
+                    {
+                        DivideAdvSimd(a, b, s);
+                        return;
+                    }
+                }
+            }
+
+            DivideNaive(a, b, s, 0, a.Length);
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideNaive(ReadOnlySpan<float> a, float b, Span<float> s, int start, int end)
+        {
+            for (var i = start; i < end; i++)
+            {
+                s[i] = a[i] / b;
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideSse(ReadOnlySpan<float> a, float b, Span<float> s)
+        {
+            var remainder = a.Length & 3;
+            var length = a.Length - remainder;
+
+            var scalar = Vector128.Create(b);
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrS = s)
+                {
+                    for (var i = 0; i < length; i += 4)
+                    {
+                        var j = Sse.LoadVector128(ptr + i);
+
+                        Sse.Store(ptrS + i, Sse.Divide(j, scalar));
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideAdvSimd(ReadOnlySpan<float> a, float b, Span<float> s)
+        {
+            var remainder = a.Length & 3;
+            var length = a.Length - remainder;
+
+            var scalar = Vector128.Create(b);
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrS = s)
+                {
+                    for (var i = 0; i < length; i += 4)
+                    {
+                        var j = AdvSimd.LoadVector128(ptr + i);
+
+                        AdvSimd.Store(ptrS + i, AdvSimd.Arm64.Divide(j, scalar));
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
+        private static void DivideAvx(ReadOnlySpan<float> a, float b, Span<float> s)
+        {
+            var remainder = a.Length & 7;
+            var length = a.Length - remainder;
+
+            var scalar = Vector256.Create(b);
+
+            fixed (float* ptr = a)
+            {
+                fixed (float* ptrS = s)
+                {
+                    for (var i = 0; i < length; i += 8)
+                    {
+                        var j = Avx.LoadVector256(ptr + i);
+
+                        Avx.Store(ptrS + i, Avx.Divide(j, scalar));
+                    }
+                }
+            }
+
+            if(remainder != 0)
+            {
+                DivideNaive(a, b, s, length, a.Length);
+            }
+        }
+
+        #endregion
+    }
+}
